\name{jvcoords-package}
\alias{jvcoords-package}
\docType{package}
\title{Package overview}
\description{
  The jvcoords package provides functions to standardize and whiten
  data, and an implementation of Principal Component Analysis (PCA).
  All three transformations are implemented using a common class
  \verb{coords} which allows to easily convert data from and to the
  new coordinate systems.

  See the documentation for \code{\link{standardize}}, \code{\link{whiten}},
  and \code{\link{PCA}} for information on how to use this package.
}
\author{
  Jochen Voss <voss@seehuhn.de>
}
\keyword{math}
\keyword{array}
\keyword{package}
\concept{multivariate}
\seealso{
  \code{\link{standardize}},
  \code{\link{whiten}},
  \code{\link{PCA}},
  \code{\link{coords}}
}
