\name{jvmulti-package}
\alias{jvmulti-package}
\docType{package}
\title{
  \packageTitle{jvmulti}
}
\description{
  The jvmulti package provides functions to standardize and whiten
  data, and an implementation of Principal Component Analysis (PCA).
  All three transformations are implemented as a common class
  \verb{coords} which allows to easily convert data from and to the
  new coordinate system.

  See the documentation for \code{standardize()}, \code{whiten()},
  and \code{jvPCA()} for information on how to use this package.
}
\author{
  Jochen Voss <voss@seehuhn.de>
}
\keyword{
  package
  multivariate
}
\seealso{
  \code{\link{standardize}},
  \code{\link{whiten}},
  \code{\link{jvPCA}},
  \code{\link{coords}}
}
